premiere implementation de HyperLogLog

+++ src/main/java/com/nedjar/estimateurs/Estimateur.java
@@ -1,5 +1,5 @@
 package com.nedjar.estimateurs;
 
 public interface Estimateur<T  extends Iterable<?>> {
-	int estimerCardinalite( T dataSet);
+	double estimerCardinalite( T dataSet);
 }
